requirements.txt
# ------------------------------------------
# MLB Player-Performance Pipeline Requirements
# Python Version: 3.10.13
# Compatible with Apple Silicon (M1/M2/M3)
# ------------------------------------------

# 🧩 Core Config + CLI
pydantic==2.6.4
python-dotenv==1.1.0
typer[all]==0.12.3
click==8.1.7
python-dateutil==2.9.0

# Config and environment
PyYAML==6.0.2
python-dotenv==1.1.0

# 📊 Data Handling
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.2
openpyxl==3.1.2

# 🌐 API Access
requests==2.31.0
git+https://github.com/toddrob99/MLB-StatsAPI.git
pybaseball==2.2.4  # ⬅️ use 2.2.4 (2.2.2 is not Py3.10-compatible)

# 🤖 Modeling
scikit-learn==1.4.2
joblib==1.4.2
xgboost==2.0.3

# 🧠 LLM / Prompt Templates
jinja2==3.1.4
httpx==0.27.0
tqdm==4.66.2

# 🧪 Testing
pytest==8.2.2
pytest-mock==3.14.0

# 🔒 Atomic File Safety (Optional)
filelock==3.14.0